исправляет функции и обработчики
