+++ src/components/offer-list/offer-list.tsx
@@ -9,13 +9,10 @@ type OfferListProps = {
 const OfferList = ({offers}: OfferListProps): JSX.Element => {
   const [activeOffer, setActiveOffer] = useState<OfferPreviewType['id'] | null>(null);
 
-  function handleCardHover(offerId: OfferPreviewType['id'] | null) {
-    setActiveOffer(offerId);
-  }
   return(
     <div className="cities__places-list places__list tabs__content">
       {offers.map((offer) => (
-        <Card key={offer.id} offer={offer} block="cities" onCardHover={handleCardHover}/>
+        <Card key={offer.id} offer={offer} block="cities" onCardHover={setActiveOffer}/>
       ))}
       {activeOffer}
     </div>
